Implement P2P HTTP and Socket Communication
- Add outbound HTTP helpers for GET and POST requests with authentication.
- Implement an HTTPS server to handle proxy, manifest, XMP, and preview requests.
- Create a local file index to manage and resolve paths for raw and proxy files.
- Introduce socket communication for event broadcasting and handling various commands.
- Define message types and structures for XMP synchronization and proxy announcements.
- Ensure persistence of local-to-remote path mappings across daemon restarts.
